Added CardDetailsHeader component
Added a new component to optimize the reutilization. Now the whole information that will be displayed in the details header will be processed by CardDetailsHeader component. That component receive three props, cardData, that is the whole card data object, subtitle, that is the subtitle that will be displayed in details header and additionalData, that is an array of strings that contains the path of the information that will be displayed also in the card details header.