Added handleError
Added actions and reducer to handle errors in the application, generating error object to render the Message component. Created Message component to display all the types of messages that will be necessary, like success, warn, error, and default messages. Removed action creators getDataByIdFailed and getDataByPageFailed, and the responsibility of handle with errors changed to handleErrorReducer. Added message colors variables to global.css.